Moves SequenceNumberWriter so it can be used by bfd-pipeline-ccw-rif.
