+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/sink/concurrent/ConcurrentRdaSink.java
@@ -3,6 +3,7 @@ package gov.cms.bfd.pipeline.rda.grpc.sink.concurrent;
 import gov.cms.bfd.pipeline.rda.grpc.ProcessingException;
 import gov.cms.bfd.pipeline.rda.grpc.RdaSink;
 import gov.cms.bfd.pipeline.sharedutils.MultiCloser;
+import gov.cms.bfd.pipeline.sharedutils.SequenceNumberTracker;
 import gov.cms.model.dsl.codegen.library.DataTransformer;
 import java.io.Closeable;
 import java.io.IOException;
