+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/sink/concurrent/SequenceNumberWriter.java
@@ -3,13 +3,15 @@ package gov.cms.bfd.pipeline.rda.grpc.sink.concurrent;
 import gov.cms.bfd.pipeline.rda.grpc.RdaSink;
 import gov.cms.bfd.pipeline.sharedutils.MultiCloser;
 import javax.annotation.concurrent.ThreadSafe;
+
+import gov.cms.bfd.pipeline.sharedutils.SequenceNumberTracker;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
 
 /**
  * An object responsible for updating the progress table in the database with the appropriate
  * sequence number as claims are written to the database. Sequence numbers are provided by calls to
- * a {@link gov.cms.bfd.pipeline.rda.grpc.sink.concurrent.SequenceNumberTracker} and updates are
+ * a {@link SequenceNumberTracker} and updates are
  * written using a {@link RdaSink}.
  *
  * @param <TMessage> type of RDA API gRPC stub object corresponding to a message
@@ -34,7 +36,7 @@ class SequenceNumberWriter<TMessage, TClaim> {
 
   /**
    * Create an instance using the provided {@link RdaSink} and {@link
-   * gov.cms.bfd.pipeline.rda.grpc.sink.concurrent.SequenceNumberTracker}.
+   * SequenceNumberTracker}.
    *
    * @param sink used to write to the database
    * @param sequenceNumbers used to track sequence number changes
@@ -46,7 +48,7 @@ class SequenceNumberWriter<TMessage, TClaim> {
 
   /**
    * Updates the sequence number in the database. Uses the current value from the {@link
-   * gov.cms.bfd.pipeline.rda.grpc.sink.concurrent.SequenceNumberTracker}. If the value has not
+   * SequenceNumberTracker}. If the value has not
    * changed nothing is written and the returned {@link Mono} will have no value. If the value has
    * changed the new value will be written to the database and emitted by the {@link Mono}. If the
    * write false the {@link Mono} will emit the {@link Exception} associated with the failure.
