+++ apps/bfd-pipeline/bfd-pipeline-shared-utils/src/test/java/gov/cms/bfd/pipeline/sharedutils/SequenceNumberTrackerTest.java
@@ -1,5 +1,6 @@
-package gov.cms.bfd.pipeline.rda.grpc.sink.concurrent;
+package gov.cms.bfd.pipeline.sharedutils;
 
+import gov.cms.bfd.pipeline.sharedutils.SequenceNumberTracker;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
